ml-system/requirements.txt
# Core ML and Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
tensorflow==2.13.0
xgboost==1.7.6
lightgbm==4.0.0

# Data Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0

# Web Framework
flask==2.3.2
flask-cors==4.0.0
gunicorn==21.2.0

# Database
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# Redis Cache
redis==4.6.0

# Background Tasks
celery==5.3.1

# Development Tools
jupyter==1.0.0
notebook==7.0.0
ipykernel==6.25.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Production
uwsgi==2.0.22